Spring first push

+++ src/main/java/com/soma/gameboy/dao/MemberDao.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.soma.gameboy.entities.Member;
 
-public interface MemberDao extends JpaRepository<Member, Long> {
+public interface MemberDao extends JpaRepository<Member, Integer> {
+
+	
 
 }
